Error is not displayed for functions with no return values.
Fixes #2749
