+++ ui/address/contract/methods/form/ContractMethodForm.tsx
@@ -19,6 +19,7 @@ import ContractMethodFieldAccordion from './ContractMethodFieldAccordion';
 import ContractMethodFieldInput from './ContractMethodFieldInput';
 import ContractMethodFieldInputArray from './ContractMethodFieldInputArray';
 import ContractMethodFieldInputTuple from './ContractMethodFieldInputTuple';
+import ContractMethodOutput from './ContractMethodOutput';
 import ContractMethodResultPublicClient from './ContractMethodResultPublicClient';
 import ContractMethodResultWalletClient from './ContractMethodResultWalletClient';
 import { getFieldLabel, matchArray, transformFormDataToMethodArgs } from './utils';
@@ -306,8 +307,13 @@ const ContractMethodForm = ({ data, attempt, onSubmit, onReset, isOpen }: Props)
           onSettle={ handleResultSettle }
         />
       ) }
-      { 'outputs' in data && data.outputs.length > 0 && (
+      { result && result.source === 'public_client' && result.data instanceof Error && (
         <ContractMethodResultPublicClient
+          data={ result.data }
+        />
+      ) }
+      { 'outputs' in data && data.outputs.length > 0 && (
+        <ContractMethodOutput
           data={ result && result.source === 'public_client' ? result.data : undefined }
           onSettle={ handleResultSettle }
           abiItem={ data }
